+++ src/Effects/Writer.hs
@@ -15,7 +15,7 @@ module Effects.Writer (
   , handleWriterM) where
 
 import Prog ( discharge, Member(inj), Prog(..) )
-import Model ( Model(..) )
+import Model ( GenModel(..) )
 
 -- | Writer effect for writing to a strean @w@
 data Writer w a where
@@ -28,9 +28,9 @@ data Writer w a where
 tell :: Member (Writer w) es => w -> Prog es ()
 tell w = Op (inj $ Tell w) Val
 
--- | Wrapper for @Tell@ inside @Model@
-tellM :: Member (Writer w) es => w -> Model env es ()
-tellM w = Model $ tell w
+-- | Wrapper for @Tell@ inside @GenModel@
+tellM :: Member (Writer w) es => w -> GenModel env es ()
+tellM w = GenModel $ tell w
 
 -- | Handle the @Writer@ effect for a stream @w@
 handleWriter :: forall w es a. Monoid w
@@ -44,9 +44,9 @@ handleWriter = loop mempty where
     Right (Tell w') -> loop (w `mappend` w') (k ())
     Left u'         -> Op u' (loop w . k)
 
--- | Handle the @Writer@ effect inside a @Model@
+-- | Handle the @Writer@ effect inside a @GenModel@
 handleWriterM :: Monoid w
-  => Model env (Writer w : es) a
+  => GenModel env (Writer w : es) a
   -- | (output, final stream)
-  -> Model env es (a, w)
-handleWriterM m = Model $ handleWriter $ runModel m
+  -> GenModel env es (a, w)
+handleWriterM m = GenModel $ handleWriter $ runModel m