+++ src/Inference/MC/PIM.hs
@@ -34,7 +34,7 @@ import qualified Inference.MC.IM as IM
 pim :: forall env vars a. (env `ContainsVars` vars)
   => Int                                            -- ^ number of MH steps
   -> Int                                            -- ^ number of particles
-  -> Model env [EnvRW env, Dist, Sampler] a                  -- ^ model
+  -> GenModel env [EnvRW env, Dist, Sampler] a                  -- ^ model
   -> Env env                                        -- ^ input environment
   -> Vars vars                                      -- ^ parameter names
   -> Sampler [Env env]                              -- ^ output environments