+++ src/Inference/MC/RMSMC.hs
@@ -62,7 +62,7 @@ pack (α, ps, τs) = zipWith3 PrtState (repeat α) ps τs
 rmsmc :: forall env a xs. (env `ContainsVars` xs)
   => Int                                          -- ^ number of SMC particles
   -> Int                                          -- ^ number of MH (rejuvenation) steps
-  -> Model env [EnvRW env, Dist, Sampler] a                -- ^ model
+  -> GenModel env [EnvRW env, Dist, Sampler] a                -- ^ model
   -> Env env                                      -- ^ input model environment
   -> Vars xs                                      -- ^ optional observable variable names of interest
   -> Sampler [Env env]                            -- ^ output model environments of each particle
@@ -79,7 +79,7 @@ rmpfilter :: es ~ '[Sampler] =>
      Int                                          -- ^ number of SMC particles
   -> Int                                          -- ^ number of MH (rejuvenation) steps
   -> [Tag]                                        -- ^ tags indicating variables of interest
-  -> ProbProg es a                                   -- ^ probabilistic program
+  -> Model es a                                   -- ^ probabilistic program
   -> Sampler [(a, PrtState)]                      -- ^ final particle results and contexts
 rmpfilter n_prts mh_steps tags model = do
   -- let q =  pfilter handleParticle model (prts, ps)
@@ -100,7 +100,7 @@ handleParticle = fmap asPrtTrace . handleM . reuseSamples Map.empty . suspendα 
 handleResample :: (es ~ '[Sampler], Member Sampler fs)
   => Int                                          -- ^ number of MH (rejuvenation) steps
   -> [Tag]                                        -- ^ tags indicating variables of interest
-  -> ProbProg es a
+  -> Model es a
   -> Handler (Resample PrtState) fs [(a, PrtState)] [(a, PrtState)]
 handleResample mh_steps tags m = handle () (const Val) (const hop) where
   hop :: Member Sampler fs => Resample PrtState x -> (() -> x -> Prog fs a) -> Prog fs a
@@ -137,8 +137,8 @@ suspendα (Op op k) = case discharge op of
 -}
 suspendAt ::
      Addr       -- ^ Address of @Observe@ operation to break at
-  -> ProbProg es a
-  -> ProbProg es (ProbProg es a)
+  -> Model es a
+  -> Model es (Model es a)
 suspendAt α_break (Val x) = pure (Val x)
 suspendAt α_break (Op op k) = case prj op of
   Just (Observe d y α) -> do