+++ src/Model.hs
@@ -13,8 +13,8 @@
 -}
 
 module Model (
-    Model(..)
-  , ProbProg
+    GenModel(..)
+  , Model
   , handleCore
     -- * Distribution smart constructors
     -- $Smart-Constructors
@@ -72,32 +72,32 @@ import Data.Typeable
     A model initially consists of (at least) two effects: @Dist@ for calling primitive distributions
     and @EnvRW env@ for reading from @env@.
 -}
-newtype Model env es a =
-  Model { runModel :: ( Member Dist es        -- models can call primitive distributions
+newtype GenModel env es a =
+  GenModel { runModel :: ( Member Dist es        -- models can call primitive distributions
                       , Member (EnvRW env) es -- models can read observed values from their environment
                       )
                    => Prog es a }
   deriving Functor
 
-instance Applicative (Model env es) where
-  pure x = Model $ pure x
+instance Applicative (GenModel env es) where
+  pure x = GenModel $ pure x
   (<*>) = ap
 
-instance Monad (Model env es) where
+instance Monad (GenModel env es) where
   return = pure
-  Model f >>= x = Model $ do
+  GenModel f >>= x = GenModel $ do
     f' <- f
     runModel $ x f'
 
 {- | Probabilistic programs are those with effects for conditioning and sampling.
 -}
 
-type ProbProg es a = Prog (Observe : Sample : es) a
+type Model es a = Prog (Observe : Sample : es) a
 
 {- | The initial handler for models, specialising a model under a certain environment
      to produce a probabilistic program consisting of @Sample@ and @Observe@ operations.
 -}
-handleCore :: Env env -> Model env (EnvRW env : Dist : es) a
+handleCore :: Env env -> GenModel env (EnvRW env : Dist : es) a
            -> Prog (Observe : Sample : es) (a, Env env)
 handleCore env_in m = (handleDist . handleEnvRW env_in) (runModel m)
 
@@ -111,7 +111,7 @@ handleCore env_in m = (handleDist . handleEnvRW env_in) (runModel m)
     variable to be conditioned against:
 
     @
-    exampleModel :: Observable env "b" Bool => Model env es Bool
+    exampleModel :: Observable env "b" Bool => GenModel env es Bool
     exampleModel = bernoulli 0.5 #b
     @
 
@@ -119,64 +119,64 @@ handleCore env_in m = (handleDist . handleEnvRW env_in) (runModel m)
     this will always representing sampling from that distribution:
 
     @
-    exampleModel' :: Model env es Bool
+    exampleModel' :: GenModel env es Bool
     exampleModel' = bernoulli' 0.5
     @
 -}
 
-callDist :: forall env x d a es. (PrimDist d a,  Show a) => Observable env x a => d -> Var x -> Model env es a
-callDist d field = Model $ do
+callDist :: forall env x d a es. (PrimDist d a,  Show a) => Observable env x a => d -> Var x -> GenModel env es a
+callDist d field = GenModel $ do
   let tag =  Just $ varToStr field
   maybe_y <- call (Read @env field)
   y <- call (Dist d maybe_y tag)
   call (Write @env field y)
   pure y
 
-callDist' :: (PrimDist d a) => d -> Model env es a
-callDist' d = Model $ call (Dist d Nothing Nothing)
+callDist' :: (PrimDist d a) => d -> GenModel env es a
+callDist' d = GenModel $ call (Dist d Nothing Nothing)
 
 dirichlet :: (Observable env x (Vec n Double), TypeableSNatI n) =>
      Vec n Double
   -> Var x
-  -> Model env es (Vec n Double)
+  -> GenModel env es (Vec n Double)
 dirichlet xs = callDist (mkDirichlet xs)
 
 dirichlet' :: (TypeableSNatI n) =>
   -- | concentration parameters
      Vec n Double
-  -> Model env es (Vec n Double)
+  -> GenModel env es (Vec n Double)
 dirichlet' xs = callDist' (mkDirichlet xs)
 
 discrete :: (Typeable a, Eq a, Show a, Observable env x a) =>
      [(a, Double)]
   -> Var x
-  -> Model env es a
+  -> GenModel env es a
 discrete ps = callDist (mkDiscrete ps)
 
 discrete' :: (Typeable a, Eq a, Show a) =>
   -- | primitive values and their probabilities
      [(a, Double)]
-  -> Model env es a
+  -> GenModel env es a
 discrete' ps = callDist' (mkDiscrete ps)
 
 categorical :: (Observable env x Int)
   => [Double]
   -> Var x
-  -> Model env es Int
+  -> GenModel env es Int
 categorical xs = callDist (mkCategorical xs)
 
 categorical' ::
   -- | list of @n@ probabilities
      [Double]
   -- | integer index from @0@ to @n - 1@
-  -> Model env es Int
+  -> GenModel env es Int
 categorical' xs = callDist' (mkCategorical xs)
 
 normal :: Observable env x Double =>
      Double
   -> Double
   -> Var x
-  -> Model env es Double
+  -> GenModel env es Double
 normal mu sigma = callDist (mkNormal mu sigma)
 
 normal'
@@ -184,26 +184,26 @@ normal'
   :: Double
   -- | standard deviation
   -> Double
-  -> Model env es Double
+  -> GenModel env es Double
 normal' mu sigma = callDist' (mkNormal mu sigma)
 
 halfNormal :: Observable env x Double
   => Double
   -> Var x
-  -> Model env es Double
+  -> GenModel env es Double
 halfNormal sigma = callDist (mkHalfNormal sigma)
 
 halfNormal'
   -- | standard deviation
   :: Double
-  -> Model env es Double
+  -> GenModel env es Double
 halfNormal' sigma = callDist' (mkHalfNormal sigma)
 
 cauchy :: Observable env x Double =>
      Double
   -> Double
   -> Var x
-  -> Model env es Double
+  -> GenModel env es Double
 cauchy mu sigma = callDist (mkCauchy mu sigma)
 
 cauchy'
@@ -211,38 +211,38 @@ cauchy'
   :: Double
   -- | scale
   -> Double
-  -> Model env es Double
+  -> GenModel env es Double
 cauchy' mu sigma = callDist' (mkCauchy mu sigma)
 
 halfCauchy :: Observable env x Double =>
      Double
   -> Var x
-  -> Model env es Double
+  -> GenModel env es Double
 halfCauchy sigma = callDist (mkHalfCauchy sigma)
 
 halfCauchy' ::
   -- | scale
      Double
-  -> Model env es Double
+  -> GenModel env es Double
 halfCauchy' sigma = callDist' (mkHalfCauchy sigma)
 
 bernoulli :: Observable env x Bool =>
      Double
   -> Var x
-  -> Model env es Bool
+  -> GenModel env es Bool
 bernoulli p = callDist (mkBernoulli p)
 
 bernoulli' ::
   -- | probability of @True@
      Double
-  -> Model env es Bool
+  -> GenModel env es Bool
 bernoulli' p = callDist' (mkBernoulli p)
 
 beta :: Observable env x Double =>
      Double
   -> Double
   -> Var x
-  -> Model env es Double
+  -> GenModel env es Double
 beta α β = callDist (mkBeta α β)
 
 beta' ::
@@ -250,14 +250,14 @@ beta' ::
      Double
   -- | shape 2 (β)
   -> Double
-  -> Model env es Double
+  -> GenModel env es Double
 beta' α β = callDist' (mkBeta α β)
 
 binomial :: Observable env x Int =>
      Int
   -> Double
   -> Var x
-  -> Model env es Int
+  -> GenModel env es Int
 binomial n p = callDist (mkBinomial n p)
 
 binomial' ::
@@ -266,14 +266,14 @@ binomial' ::
   -- | probability of successful trial
   -> Double
   -- | number of successful trials
-  -> Model env es Int
+  -> GenModel env es Int
 binomial' n p = callDist' (mkBinomial n p)
 
 gamma :: forall env es x. Observable env x Double =>
      Double
   -> Double
   -> Var x
-  -> Model env es Double
+  -> GenModel env es Double
 gamma k θ = callDist (mkGamma k θ)
 
 gamma' ::
@@ -281,14 +281,14 @@ gamma' ::
      Double
   -- | scale (θ)
   -> Double
-  -> Model env es Double
+  -> GenModel env es Double
 gamma' k θ = callDist' (mkGamma k θ)
 
 uniform :: Observable env x Double =>
      Double
   -> Double
   -> Var x
-  -> Model env es Double
+  -> GenModel env es Double
 uniform min max = callDist (mkUniform min max)
 
 uniform' ::
@@ -296,19 +296,19 @@ uniform' ::
      Double
   -- | upper-bound
   -> Double
-  -> Model env es Double
+  -> GenModel env es Double
 uniform' min max = callDist' (mkUniform min max)
 
 poisson :: Observable env x Int =>
      Double
   -> Var x
-  -> Model env es Int
+  -> GenModel env es Int
 poisson λ = callDist (mkPoisson λ)
 
 poisson' ::
   -- | rate (λ)
      Double
   -- | number of events
-  -> Model env es Int
+  -> GenModel env es Int
 poisson' λ = callDist' (mkPoisson λ)